[core-kit] media-sound/alsa-plugins Refactor and update the media-sound/alsa-plugins in the core-kit
This commit includes several changes to the media-sound/alsa-plugins within the core-kit. A significant part of the update is the removal of the FFMPEG_SUBSLOT in the ffmpeg template file, and changing the value in the SLOT variable from 0/${FFMPEG_SUBSLOT} to just 0.

The version of alsa-lib and alsa-plugins in the autogen.yaml file was also updated to 1.2.13 and 1.2.12 respectively. Furthermore, the patch in the alsa-lib template file for the 652422 bugfix was removed, as it is no longer necessary.

File paths in the alsa-plugins template were updated, changing from {{gen_path}}/files/${PN}/ to "${FILESDIR}/". This results in cleaner and verifiable paths for the pulse-default.conf and 51-pulseaudio-probe.conf files.

Lastly, in-house files like the alsa-lib-1.1.6-missing_files.patch were deleted, improving the overall cleanliness and organization of the codebase.